Fixed
This version is now working again on the server. The problem is
including certain files that will lead to compilation errors. See the
CG.hpp for details.

Also the Algorithm can now terminate earlier if a certain error
threshold has been reached. This is massiveley increasing the runtime
with virtually no impact on accuary, which is surprising to me.